LICENSE
Copyright (c) 2025 Angel Silva  All rights reserved.  This software and its source code are the property of Angel Silva.  Permission is granted to individuals to access and use this app via its published website for personal, non-commercial use only.  You may not: - Copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the   source code or any substantial portion of it; - Repost or repackage this app or its code as your own product; - Use the code in other projects without explicit written permission from the owner.  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.